Showing 50 nearest cities to current location.

+++ app/src/main/java/com/android/forecasty/data/model/town/Coord.kt
@@ -2,5 +2,5 @@ package com.android.forecasty.data.model.town
 
 import com.google.gson.annotations.SerializedName
 
-data class Coord(@SerializedName("lon") val lon: Int,
-                 @SerializedName("lat") val lat: Int)
+data class Coord(@SerializedName("lon") val lon: Double,
+                 @SerializedName("lat") val lat: Double)
